# games/docs/api.yaml
openapi: 3.0.3
info:
  title: Soleo game creator API
  description: Soleo game creator API
  termsOfService: https://soleo.games/
  contact:
    email: djura@soleo.games
    name: Djuradj Gvozdenovic
  version: 1.0.1
externalDocs:
  description: Soleo games website
  url: https://soleo.games/
servers:
  - url: http://localhost:8080/api/v1
tags:
  - name: question
    description: Manage Questions
    externalDocs:
      description: Find out more
      url: http://swagger.io
  - name: task
    description: Manage Tasks
    externalDocs:
      description: Find out more
      url: http://swagger.io
paths:
  /questions:
    post:
      tags:
        - question
      summary: Create new question
      description: Create new question
      operationId: createQuestion
      requestBody:
        description: Create new question
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Question'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
        '409':
          description: conflict
    get:
      tags:
        - question
      summary: Search questions
      description: Search questions
      operationId: searchQuestion
      parameters:
        - name: category
          in: query
          description: Category to filter by
          required: false
          schema:
            type: string
            example: sport
        - name: tags
          in: query
          description: Tags to filter by
          required: false
          explode: true
          schema:
            type: array
            items:
              type: string
        - name: difficulty
          in: query
          description: Difficulty to filter by
          required: false
          schema:
            type: string
            example: easy
        - name: language
          in: query
          description: Language to filter by
          required: false
          schema:
            type: string
            example: english
        - name: type
          in: query
          description: Question type to filter by
          required: false
          schema:
            type: string
            example: corporate
        - name: limit
          in: query
          description: Limit for the number of questiones returned
          required: false
          schema:
            type: integer
            example: 100
        - name: offset
          in: query
          description: Offset for pagination
          required: false
          schema:
            type: integer
            example: 0
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    description: Total number of questions available
                    example: 100
                  limit:
                    type: integer
                    description: Number of questions returned in this response
                    example: 10
                  offset:
                    type: integer
                    description: Offset used for pagination
                    example: 0
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Question'
            text/csv:
              schema:
                type: string
                example: |
                  questionText,answerText,category,difficulty,type
                  Which is the most trophy football club in Macedonia {player name}?,FK Vardar,sport,easy,corporate

        '400':
          description: Invalid input
        '422':
          description: Validation exception
        '409':
          description: conflict
  /questions/import:
    post:
      tags:
        - question
      summary: Import questions
      description: Import questions
      operationId: importQuestions
      requestBody:
        description: Import questions
        content:
          text/csv:
            schema:
              type: string
              example: |
                questionText,answerText,category,difficulty,type
                Which is the most trophy football club in Macedonia {player name}?,FK Vardar,sport,easy,corporate
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
        '409':
          description: conflict
  /tasks:
    post:
      tags:
        - task
      summary: Create new task
      description: Create new task
      operationId: createTask
      requestBody:
        description: Create new task
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
        '409':
          description: conflict
    get:
      tags:
        - task
      summary: Search tasks
      description: Search tasks
      operationId: searchTasks
      parameters:
        - name: category
          in: query
          description: Category to filter by
          required: false
          schema:
            type: string
            example: sport
        - name: tags
          in: query
          description: Tags to filter by
          required: false
          explode: true
          schema:
            type: array
            items:
              type: string
        - name: difficulty
          in: query
          description: Difficulty to filter by
          required: false
          schema:
            type: string
            example: easy
        - name: language
          in: query
          description: Language to filter by
          required: false
          schema:
            type: string
            example: english
        - name: type
          in: query
          description: Task type to filter by
          required: false
          schema:
            type: string
            example: corporate
        - name: limit
          in: query
          description: Limit for the number of tasks returned
          required: false
          schema:
            type: integer
            example: 100
        - name: offset
          in: query
          description: Offset for pagination
          required: false
          schema:
            type: integer
            example: 0
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    description: Total number of tasks available
                    example: 100
                  limit:
                    type: integer
                    description: Number of tasks returned in this response
                    example: 10
                  offset:
                    type: integer
                    description: Offset used for pagination
                    example: 0
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
            text/csv:
              schema:
                type: string
                example: |
                  text,category,difficulty,type
                  Do something,corporate,easy,sport

        '400':
          description: Invalid input
        '422':
          description: Validation exception
        '409':
          description: conflict
  /tasks/import:
    post:
      tags:
        - task
      summary: Import task
      description: Import task
      operationId: importTask
      requestBody:
        description: Import task
        content:
          text/csv:
            schema:
              type: string
              example: |
                text,category,difficulty,type
                Do something,corporate,easy,sport
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
        '409':
          description: conflict
components:
  schemas:
    Question:
      required:
        - questionText
        - category
        - tags
        - difficulty
        - type
      type: object
      properties:
        questionText:
          type: string
          example: Which is the most trophy football club in Macedonia, {player name}?
        category:
          type: string
          example: sport
        tags:
          type: array
          items:
            type: string
            example: Macedonia
        difficulty:
          type: string
          example: easy
        type:
          type: string
          example: corporate
        answerText:
          type: string
          example: FK Vardar
        questionTranslations:
          type: array
          items:
            $ref: '#/components/schemas/Translation'
          example: [{"language":"Español", "value":"¿Cuál es el club de fútbol con más trofeos de Macedonia, {player name}?"}]
        answerTranslations:
          type: array
          items:
            $ref: '#/components/schemas/Translation'
          example: [{"language":"Español", "value":"FK Vardar"}]
        parameters:
          type: array
          items:
            $ref: '#/components/schemas/Parameter'
    Task:
      required:
        - text
        - category
        - tags
        - difficulty
        - type
      type: object
      properties:
        text:
          type: string
          example: Do something
        category:
          type: string
          example: office
        tags:
          type: array
          items:
            type: string
            example: Macedonia
        difficulty:
          type: string
          example: easy
        type:
          type: string
          example: corporate
        translations:
          type: array
          items:
            $ref: '#/components/schemas/Translation'
          example: [{"language":"Español", "value":"hacer algo"}]
    Translation:
      required:
        - language
        - value
      type: object
      properties:
        language:
          type: string
        value:
          type: string
    Parameter:
      required:
        - name
        - type
      type: object
      properties:
        name:
          type: string
          example: player name
        type:
          type: string
          example: string
        description:
          type: string
          example: name of the main player
      